requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.0.0